lifetouch-dwh-data-pipeline-airflow-local/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.11.1
anyio==3.7.1
apache-airflow==2.6.3
apache-airflow-providers-amazon==8.15.0
apache-airflow-providers-common-sql==1.5.2
apache-airflow-providers-databricks==4.7.0
apache-airflow-providers-ftp==3.4.2
apache-airflow-providers-http==4.4.2
apache-airflow-providers-imap==3.2.2
apache-airflow-providers-oracle==3.7.1
apache-airflow-providers-postgres==5.7.0
apache-airflow-providers-sqlite==3.4.2
apispec==5.2.2
argcomplete==3.1.1
asgiref==3.7.2
asn1crypto==1.5.1
astro-provider-databricks==0.1.6
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
beautifulsoup4==4.12.3
black==23.12.1
blinker==1.6.2
boto3==1.34.21
botocore==1.34.21
cachelib==0.9.0
cattrs==23.1.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==40.0.2
databricks==0.2
databricks-cli==0.18.0
databricks-sql-connector==2.9.3
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.3.0
docutils==0.20.1
email-validator==1.3.1
et-xmlfile==1.1.0
exceptiongroup==1.1.2
flake8==7.0.0
Flask==2.2.5
Flask-AppBuilder==4.3.1
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.2
Flask-Limiter==3.3.1
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.1.1
frozenlist==1.3.3
google-re2==1.0
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-resources==5.12.0
inflection==0.5.1
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpath-ng==1.6.1
jsonschema==4.18.0
jsonschema-specifications==2023.6.1
lazy-object-proxy==1.9.0
limits==3.5.0
linkify-it-py==2.0.2
lockfile==0.12.2
lxml==5.1.0
lz4==4.3.3
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mccabe==0.7.0
mdit-py-plugins==0.4.0
mdurl==0.1.2
mergedeep==1.3.4
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.3
oauthlib==3.2.2
openpyxl==3.1.2
oracledb==2.0.1
ordered-set==4.1.0
packaging==23.2
pandas==2.1.4
pathspec==0.9.0
pendulum==2.1.2
platformdirs==4.1.0
pluggy==1.2.0
ply==3.11
prison==0.2.1
psutil==5.9.5
psycopg2-binary==2.9.9
pyarrow==14.0.2
pycodestyle==2.11.1
pycparser==2.21
pydantic==1.10.11
pyflakes==3.2.0
Pygments==2.15.1
PyJWT==2.7.0
pyparsing==3.1.0
pytest==7.4.4
pytest-mock==3.12.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3
pytzdata==2020.1
PyYAML==6.0
redshift-connector==2.0.918
referencing==0.29.1
requests==2.31.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.4.2
rich-argparse==1.2.0
rpds-py==0.8.8
s3transfer==0.10.0
scramp==1.4.4
setproctitle==1.3.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
sqlalchemy-redshift==0.8.14
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
tabulate==0.9.0
tenacity==8.2.2
termcolor==2.3.0
text-unidecode==1.3
thrift==0.16.0
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.4
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.16
watchtower==3.0.1
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yamllint==1.33.0
yarl==1.9.2
